% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{all_breakends}
\alias{all_breakends}
\title{all_breakends}
\usage{
all_breakends(jabba_rds, return.type = "GRanges", verbose = FALSE)
}
\arguments{
\item{jabba_rds}{(character) path to JaBbA graph}

\item{return.type}{(character) default GRanges}

\item{verbose}{(logical)}
}
\value{
GRanges or data.table
}
\description{
Get all breakends from JaBbA and say whether they are loose or junctions
}
