% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{anchorlift_coverage}
\alias{anchorlift_coverage}
\title{anchorlift_coverage}
\usage{
anchorlift_coverage(
  bps = NULL,
  cov_rds = NA_character_,
  normal_cov_rds = NA_character_,
  field = "foreground",
  purity = NA,
  ploidy = NA,
  return.type = "data.table",
  mask = "~/projects/gGnome/files/zc_stash/maskC.rds",
  window = 1e+05,
  ignore.strand = FALSE,
  round.digits = -3,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{bps}{(GRanges or data.table coercible to GRanges)}

\item{cov_rds}{(character) path to coverage file}

\item{normal_cov_rds}{(character) path to normal coverage file}

\item{field}{(character) coverage field, default foreground}

\item{purity}{(numeric) estimated purity}

\item{ploidy}{(numeric) estimated ploidy}

\item{return.type}{(character) default data.table}

\item{mask}{(character) path to coverage mask}

\item{window}{(numeric)  anchorlift window, default 1e5}

\item{ignore.strand}{(logical) ignore strand for anchorlift}

\item{round.digits}{(round anchorlift start for binning)}

\item{verbose}{(logical) default FALSE}

\item{...}{other inputs to anchorlift}
}
\value{
anchorlifted coverage relative to breakpoints
}
\description{
anchorlift_coverage
}
