% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{assays}
\alias{assays}
\title{Access assays}
\usage{
\S4method{assays}{Zi}(x, withDimnames = TRUE, ...)
}
\arguments{
\item{ZiObject}{'Zi'-class object}
}
\value{
list
}
\description{
access  assays of an 'Zi'-class object if the datafile is an
object of the class SummarizedExperiment
}
