% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{boxplot}
\alias{boxplot}
\alias{boxplot.Zi}
\title{Create boxplots of a 'Zi'-class object}
\usage{
\method{boxplot}{Zi}(x, log1p = FALSE, ...)
}
\arguments{
\item{x}{'Zi'-class object}

\item{log1p}{logical, default = FALSE, if TRUE log(1+p) transformation takes
place}

\item{...}{see boxplot documentation}
}
\description{
Create boxplots of a 'Zi'-class object.
}
\seealso{
\link[graphics]{boxplot}
}
