% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{breakpoint_coverage}
\alias{breakpoint_coverage}
\title{breakpoint_coverage}
\usage{
breakpoint_coverage(
  bps = GRanges(),
  cov_rds = NULL,
  field = "foreground",
  pad = 50000,
  mask = "~/projects/gGnome/files/lowmap.rds",
  verbose = FALSE
)
}
\arguments{
\item{bps}{(GRanges) breakpoints}

\item{cov_rds}{(character) coverage file}

\item{field}{(character) default foreground}

\item{pad}{(numeric) default 5e4}

\item{mask}{(character) path to coverage mask}

\item{verbose}{(logical) default FALSE}
}
\value{
data.table with fields:
- cn (from coverage)
- bp.index (index from breakpoints)
}
\description{
Get coverage bins within a certain distance of fused and unfused sides of breakpoints
}
