% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcWeights}
\alias{calcWeights}
\title{calculate weights}
\usage{
calcWeights(zi_input, zi, feature, dist = c("poisson", "negbin", "geometric"))
}
\arguments{
\item{zi_input}{result of reshape_zi, count data in a long format}

\item{zi}{result of fitting a zero inflation model to the data using pscl::zeroinfl}

\item{feature}{character string characterizing the rows, e.g. gene, OTU, ...}
}
\value{
matrix of the calculated weights
}
\description{
the posterior probability that a given count y arises from the
NB count component can be used as weights and is calculated using the Bayes'
rule: w = ((1-pi)*fnb)/fzinb
}
