% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{cbs_correlation}
\alias{cbs_correlation}
\title{cbs_correlation}
\usage{
cbs_correlation(
  cov = NA_character_,
  replication = "~/DB/fishHook/covariates/replication.timing.rds",
  field = "foreground",
  rep.field = "val",
  log = TRUE,
  bins.size = 1e+06
)
}
\arguments{
\item{cov}{(character) path to coverage file}

\item{replication}{(character) path to replication timing GRanges}

\item{bin.size}{(numeric) median rebin}
}
\value{
list with three elements
- data: data.table of bins with foreground, replication timing
- pearson: pearson correlation
- spearman: spearman correlation
}
\description{
cbs_correlation
}
