% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{check_class}
\alias{check_class}
\title{check_class}
\usage{
check_class(obj = NULL, class = character())
}
\arguments{
\item{class}{(character)}
}
\value{
logical, TRUE if obj is an instance of class
}
\description{
Test if object is an instance of some class
But also if object is null/NA, which breaks inherits()
}
