% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{check_loose_recovery}
\alias{check_loose_recovery}
\title{check_loose_recovery}
\usage{
check_loose_recovery(
  jabba_rds = "/dev/null",
  le_dt = "~/projects/lambda/db/le.rds",
  id = "",
  pad = 10000,
  mask = "~/projects/gGnome/files/zc_stash/maskA_re.rds",
  return.type = "data.table",
  verbose = FALSE
)
}
\arguments{
\item{jabba_rds}{(path to JaBbA gGraph)}

\item{le_dt}{(path to data table)}

\item{id}{(character) id of sample}

\item{pad}{(numeric) pad for overlaps}

\item{return.type}{character, c("GRanges", "data.table")}

\item{verbose}{(logical) default FALSE}
}
\value{
loose ends with metadata of ov.junction and ov.loose and ov.mask
}
\description{
Given a JaBbA graph and set of true positive breakpoints:
check how many TP loose ends are recovered by either JaBbA junctions or loose ends
}
