% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{cluster2walks}
\alias{cluster2walks}
\title{cluster2walks}
\usage{
cluster2walks(junctions, gaps, verbose = FALSE)
}
\arguments{
\item{junctions}{(Junction) involved rearrangements}

\item{gaps}{(GRanges) signed gaps}

\item{verbose}{(logical) default FALSE}
}
\description{
get all walks associated with an edge cluster
}
