% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{cnloh_analysis}
\alias{cnloh_analysis}
\title{cnloh_analysis}
\usage{
cnloh_analysis(
  gg,
  gs,
  gg.fields = c("cn.high", "cn.low"),
  gs.fields = c("acn", "bcn"),
  max.dist = 1e+05,
  new = TRUE,
  id = "sample",
  method = "gg",
  exclude = c("X", "Y"),
  tile.width = 10000,
  mask = "~/projects/gGnome/files/zc_stash/maskA_re.rds",
  confusion = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{gg}{(character) path to gGraph under consideration}

\item{gs}{(character) path to gold standard segmentation under consideration}

\item{gg.fields}{(character) character vector with length two giving fields for the two alleles in gg}

\item{gs.fields}{(character) character vector with length two giving fields for the two alleles in gs}

\item{max.dist}{(character) max allowable distance to true positive call (default 1e5, generous)}

\item{new}{(logical) use "new" way where we benchmark based on locus overlap (default TRUE)}

\item{id}{(character) sample ID}

\item{method}{(character) input method (default gg, one of c("gr", "csv", "df", "gg"))}

\item{exclude}{(character) character vector of chromosomes to exclude}

\item{tile.width}{(numeric) tile width for considering cnloh (default 1e4)}

\item{mask}{(character) path to coverage mask}

\item{confusion}{(logical) if TRUE returns a data.table in cnloh.res that can be used to make a confusion matrix with geom_tile}

\item{verbose}{(logical) default FALSE}
}
\value{
list with two elements, cnloh.res with summary statistics and cnloh.gr with ranges of segments
}
\description{
Compare the CNLOH points in two graphs with allelic CN annotations
"CNLOH" here is defined broadly: it is any site where the CN of parental alleles change in opposite directions
However, the total CN stays the same
}
