% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{colData}
\alias{colData}
\title{Access the col Data}
\usage{
\S4method{colData}{Zi}(x, ...)
}
\arguments{
\item{ZiObject}{\code{\linkS4class{Zi}}-class object}
}
\value{
DFrame
}
\description{
access the colData of an \code{\linkS4class{Zi}}-class object if the inputdata is an
object of the class SummarizedExperiment
}
