% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{colMeans}
\alias{colMeans}
\alias{rowMeans}
\title{Calculate the row or column means of zero-inflated count data}
\usage{
\S4method{colMeans}{Zi}(x)

\S4method{rowMeans}{Zi}(x)
}
\arguments{
\item{x}{A'Zi'-class object}
}
\value{
a numeric vector of row/column length
}
\description{
Calculate row and column means of zero-inflated count data taking
weights for structural zeros into account.
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
colMeans(Zi)
rowMeans(Zi)
}
