% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{compare_cn}
\alias{compare_cn}
\title{compare_cn}
\usage{
compare_cn(gr = NULL, tiles = NULL, nm = "cn", verbose = FALSE)
}
\arguments{
\item{gr}{(GRanges) test CN}

\item{tiles}{(GRanges) tiles to evaluate CN fit over}

\item{nm}{(character) metadata field to store output, default cn}

\item{verbose}{(logical)}
}
\value{
data.table of overlapping ranges
}
\description{
compare CN of two GRanges objects, which each must have field $cn
}
