% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{compare_cncp}
\alias{compare_cncp}
\title{compare_cncp}
\usage{
compare_cncp(
  gr = NULL,
  gs = NULL,
  max.dist = 1000,
  verbose = FALSE,
  ignore.strand = FALSE
)
}
\arguments{
\item{gr}{(GRanges) test CN change points}

\item{gs}{(GRanges) gold standard CN change points}

\item{max.dist}{(numeric) max allowable distance in BP for a match}

\item{verbose}{(logical)}
}
\value{
list with cncp.res and cncp.gr
}
\description{
compare locations of CN change points given by two GRanges objects
}
