% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{compare_jabba_cn}
\alias{compare_jabba_cn}
\title{compare_jabba_cn}
\usage{
compare_jabba_cn(
  gg1 = NULL,
  gg2 = NULL,
  field1 = "cn",
  field2 = "cn",
  id = "sample",
  tile.width = 10000,
  min.width = 1000,
  verbose = FALSE
)
}
\arguments{
\item{gg1}{(character) path to JaBbA 1}

\item{gg2}{(character) path to JaBbA 2}

\item{field1}{(character) default cn}

\item{field2}{(character) default cn}

\item{id}{(character) sample ID}

\item{tile.width}{(numeric) resolution at which to tile the genome for correlation. Default 1e4.}

\item{min.width}{(numeric) min tile width to consider for RMSE, etc.}

\item{verbose}{(logical) default FALSE}
}
\value{
list with entries $cncp.gr $cncp.res
}
\description{
This is a function that compares segment CNs in two jabba outputs
}
