% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{compare_linked_junctions}
\alias{compare_linked_junctions}
\title{compare_linked_junctions}
\usage{
compare_linked_junctions(
  jj = jJ(),
  exclude = jJ(),
  jabba_rds = NA_character_,
  ascat_seg = NA_character_,
  sequenza_seg = NA_character_,
  germline_vcf = NA_character_,
  mask_gr = "~/projects/gGnome/files/zc_stash/maskA_re.rds",
  pad = 10000,
  min_size = 10000,
  std_chrs = TRUE,
  return.type = "data.table",
  keep.inv = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{jj}{(Junction) Junction object}

\item{exclude}{(Junction) Junction object with junctions to exclude}

\item{jabba_rds}{(character) path to jabba file}

\item{ascat_seg}{(character) path to ascat segmentation}

\item{sequenza_seg}{(character) path to sequenza segmentation}

\item{germline_vcf}{(character) path to germline junctions file}

\item{mask_gr}{(character) path to mask}

\item{pad}{(numeric) pad around breakends, default 1e4}

\item{min_size}{(numeric) min span of indels, default 1e4}

\item{std_chrs}{(logical) default TRUE}

\item{return.type}{(character) one of "data.table", "GRanges"}

\item{verbose}{(logical) default FALSE}
}
\value{
data.table of breakpoints with columns
- seqnames, start, end, strand (location and orientation)
- id (character) id of junction in original dataset
}
\description{
compare_linked_junctions
}
