% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R, R/plots.R
\name{cor}
\alias{cor}
\title{Calculate weighted Pearson Correlation coeffiecients}
\usage{
\S4method{cor}{Zi,ANY}(x, y = NULL, use = "everything", method = "pearson")

\S4method{cor}{Zi,ANY}(x, y = NULL, use = "everything", method = "pearson")
}
\arguments{
\item{x}{'Zi'-class object}
}
\description{
calculate the weighted pearson correlation coefficients of a count matrix
of an Zi object taking weights for structural zeros into account

calculate the weighted pearson correlation coefficients of a count matrix
of an Zi object taking weights for structural zeros into account
}
