% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cov}
\alias{cov}
\alias{cov,Zi,ANY-method}
\title{Calculate weighted Covariance}
\usage{
\S4method{cov}{Zi,ANY}(x, y = NULL, use = "everything")
}
\arguments{
\item{x}{'Zi'-class object}

\item{y}{'Zi'-class object}

\item{use}{"everything"}
}
\value{
covariance matrix
}
\description{
calculate the weighted covariance of the columns of the
count matrix of an Zi object taking weights for possible structural zero
counts into account
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
cov(Zi)
}
