% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{create_mappability_fasta}
\alias{create_mappability_fasta}
\title{create_mappability_fasta}
\usage{
create_mappability_fasta(
  ref,
  width = 101,
  gr = NULL,
  output.seqnames = c(),
  std.chrs = TRUE,
  tile = FALSE,
  output.dir = "~/testing_tmp/mappability",
  format = "fastq",
  verbose = TRUE
)
}
\arguments{
\item{ref}{(DNAStringSet or fasta)}

\item{width}{(numeric) bp, default 101}

\item{gr}{(GRanges) limit output to bases starting within these ranges}

\item{output.seqnames}{(character) limit output to these particular seqnames}

\item{std.chrs}{(logical) default TRUE}

\item{output.dir}{(character) path to output directory}

\item{format}{(character) either fasta or fastq}

\item{verbose}{(logical) print stuff}
}
\value{
(invisibly) returns path to output FASTA/FASTQ file
}
\description{
create_mappability_fasta
}
