% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{deinflatedcounts}
\alias{deinflatedcounts}
\alias{deinflatedcounts,Zi-method}
\title{Access the model}
\usage{
deinflatedcounts(x)

\S4method{deinflatedcounts}{Zi}(x)
}
\arguments{
\item{x}{\code{\linkS4class{Zi}}-class object}
}
\value{
deinflatedcounts
}
\description{
access the deinflatedcounts of an \code{\linkS4class{Zi}}-class object
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
deinflatedcounts(Zi)
}
