% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{dgv2grl}
\alias{dgv2grl}
\title{dgv2grl}
\usage{
dgv2grl(
  dgv.fn,
  colnums = numeric(),
  colnames = character(),
  mcols = c(),
  field = "varianttype",
  svtypes = c("deletion", "gain", "loss", "duplication")
)
}
\arguments{
\item{dgv.fn}{(character) path to text file}

\item{colnums}{(numeric) column numbers to extract}

\item{colnames}{(column names)}

\item{svtypes}{(character) default c("deletion", "duplication", "gain", "loss") (ignores case)}
}
\value{
GRangesList with values field
}
\description{
convert table from DGV to GRangesList
}
