% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{easy.cut}
\alias{easy.cut}
\title{easy.cut}
\usage{
easy.cut(x, start = 1000, end = -start, step = 100)
}
\arguments{
\item{x}{(numeric) vector of values}

\item{start}{(numeric) lowest number to start at, default -1e3}

\item{end}{(numeric) highest number to end at, default - start}

\item{step}{(numeric) step size, default 100}
}
\value{
numeric vector with bin midpoints
}
\description{
wrapper around cut to save literally one line of code
}
