% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{egraph}
\alias{egraph}
\title{egraph}
\usage{
egraph(
  gg,
  altedges,
  thresh = 10000,
  verbose = FALSE,
  chunksize = 1e+30,
  mc.cores = 1
)
}
\arguments{
\item{gg}{(gGraph)}

\item{thresh}{(numeric) distance threshold}

\item{verbose}{(logical) default FALSE}

\item{mc.cores}{(numeric) default 1}
}
\value{
adj (adjacency matrix of edge graph)
}
\description{
make edge graph
}
