% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{fused_unfused}
\alias{fused_unfused}
\title{fused_unfused}
\usage{
fused_unfused(loose.dt, jabba_rds = NA_character_, verbose = TRUE)
}
\arguments{
\item{loose.dt}{(data.table)}

\item{jabba_rds}{(JaBbA)}

\item{verbose}{(logical)}
}
\value{
data.table with columns:
- loose.index
- fused.node.id
- unfused.node.id
- fused.side
}
\description{
fused_unfused
}
