% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{get_consensus}
\alias{get_consensus}
\title{get_consensus}
\usage{
get_consensus(
  ascat = NA_character_,
  sequenza = NA_character_,
  breaks = "ascat",
  pad = 1e+05,
  min.width = 50000,
  max.gap = 50000,
  verbose = FALSE
)
}
\arguments{
\item{ascat}{(character) path to ascat segnemtation}

\item{sequenza}{(character) path to sequenza segmentation}

\item{breaks}{(character) 'ascat' or 'sequenza' (which breakpoints to use)}

\item{pad}{(numeric)}

\item{min.width}{(numeric) minimum width in bp for an SV default 5e4}

\item{verbose}{(logical)}
}
\value{
GRanges of consensus loose ends
}
\description{
Get consensus breakends between sequenza and ascat
}
