% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{gg2jab}
\alias{gg2jab}
\title{gg2jab}
\usage{
gg2jab(gg, purity = NA, ploidy = NA)
}
\arguments{
\item{gg}{(gGraph)}

\item{purity}{(numeric) overwrite automated purity (default 1)}

\item{ploidy}{(numeric) overwrite automated purity ploidy calculation}
}
\value{
list with names:
- segstats (GRanges, signed)
- adj (adjacency matrix)
- ab.edges (array)
- purity (numeric)
- ploidy (numeric)
- junctions (GRangesList)
}
\description{
Hacky way to make a JaBbA-like output from a gGraph
}
