% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{gg_loose_end}
\alias{gg_loose_end}
\title{gg_loose_end}
\usage{
gg_loose_end(jabba_rds = NA_character_, return.type = "GRanges")
}
\arguments{
\item{jabba_rds}{(character) path to jabba output}

\item{return.type}{(character) one of GRanges, data.table}
}
\value{
loose ends of a JaBbA graph as GRanges
if gGraph is not balanced, will produce a warning and return empty GRanges
}
\description{
gg_loose_end
}
