% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{grab.allelic.cn.gtrack}
\alias{grab.allelic.cn.gtrack}
\title{grab.allelic.cn.gtrack}
\usage{
grab.allelic.cn.gtrack(
  gg,
  melted = TRUE,
  afield = "allele",
  cfield = "cn",
  colormap = c(major = alpha("red", 0.5), minor = alpha("blue", 0.5)),
  jitter = 0.18,
  ywid = 0.3,
  ...
)
}
\arguments{
\item{gg}{(character) path to gGraph or a GRanges}

\item{melted}{(logical) default TRUE}

\item{afield}{(character) allele field (if melted is TRUE)}

\item{cfield}{(cn field) CN field (if melted is FALSE)}

\item{colormap}{(character) named character vector}

\item{jitter}{(numeric) jitter major nodes up this amount and minor nodes down this amount}

\item{ywid}{(numeric)}

\item{...}{other inputs to gTrack}
}
\value{
gTrack with desired properties
}
\description{
grab jittered gTrack showing allelic CN
}
