% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{grab.cov.gtrack}
\alias{grab.cov.gtrack}
\title{grab.cov.gtrack}
\usage{
grab.cov.gtrack(
  cov,
  jab = NULL,
  purity = NULL,
  ploidy = NULL,
  field = "ratio",
  lwd.border = 0.2,
  max.ranges = 10000,
  name = "cov",
  mask = "~/projects/gGnome/files/zc_stash/maskA_re.rds",
  chrsub = TRUE,
  hidemask = FALSE,
  ...
)
}
\arguments{
\item{cov}{(character) path to cov.rds}

\item{jab}{(character) path to jabba output}

\item{purity}{(numeric) estimated purity (overridden by JaBbA)}

\item{ploidy}{(numeric) estimated ploidy (overridden by JaBbA)}

\item{field}{(character) field in coverage file}

\item{lwd.border}{(numeric)}

\item{max.ranges}{(numeric)}

\item{name}{(character)}

\item{mask}{(character) path to coverage mask}

\item{...}{additional params for gTrack}
}
\value{
gTrack of coverage
}
\description{
create coverage gTrack
}
