% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{grab.hets.from.maf}
\alias{grab.hets.from.maf}
\title{grab.hets.from.maf}
\usage{
grab.hets.from.maf(agt.fname, min.frac = 0.2)
}
\arguments{
\item{agt.fname}{(character)}

\item{min.frac}{(numeric)}
}
\description{
get hets into format needed by phased.binstats from HMF maf_approx
}
