% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{grab.hets.gtrack}
\alias{grab.hets.gtrack}
\title{grab.hets.gtrack}
\usage{
grab.hets.gtrack(
  hets = NULL,
  jab = NULL,
  purity = NULL,
  ploidy = NULL,
  max.ranges = 5000,
  ...
)
}
\arguments{
\item{hets}{(character) variants text file}

\item{jab}{(character)}

\item{...}{additional params for gTrack}
}
\description{
create gTrack for hets
}
