% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{grab_bad_segs}
\alias{grab_bad_segs}
\title{grab_bad_segs}
\usage{
grab_bad_segs(
  jabba_rds,
  gs = NULL,
  kag = NULL,
  field1 = "cn",
  field2 = "cn",
  bad.thresh = 1,
  mask = "~/projects/gGnome/files/new.mask.rds",
  return.type = "data.table",
  verbose = FALSE
)
}
\arguments{
\item{jabba_rds}{(character) path to jabba output}

\item{gs}{(character) path to gold standard GRanges}

\item{kag}{(character) path to karyograph - compare to cn.old}

\item{field1}{(character) default cn}

\item{field2}{(character) default cn}

\item{bad.thresh}{(character) numeric, default 1}

\item{mask}{(character) path to mask}

\item{return.type}{(character) either GRanges or data.table}

\item{verbose}{(logical)}

\item{GRanges}{with columns cn, cn.gs, and diff (or data.table)}
}
\description{
Return GRanges with bad segments (high difference between fields)
}
