% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{grab_centromeres}
\alias{grab_centromeres}
\title{grab_centromeres}
\usage{
grab_centromeres(pad = 0)
}
\arguments{
\item{pad}{centromeres?}
}
\value{
GRanges of centromeres
}
\description{
grab_centromeres
}
