% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{grab_cncp}
\alias{grab_cncp}
\title{grab_cncp}
\usage{
grab_cncp(gr = NULL, include.na = TRUE, verbose = FALSE)
}
\arguments{
\item{gr}{(GRanges) with metadata field $cn}

\item{include.na}{(logical) include segments with NA CN as CN changepoint}

\item{verbose}{(logical)}
}
\value{
GRanges each with width 1 giving CN change points
}
\description{
get copy number changepoints from simplifed CN segmentation
}
