% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{grab_cnloh}
\alias{grab_cnloh}
\title{grab_cnloh}
\usage{
grab_cnloh(
  gr,
  a.field = "A",
  b.field = "B",
  ranges = TRUE,
  return.type = "GRanges",
  general = TRUE,
  balanced.only = FALSE
)
}
\arguments{
\item{gr}{(GRanges)}

\item{a.field}{(character) field of A allele CN}

\item{b.field}{(character) field of B allele CN}

\item{ranges}{(logical) return actual ranges of (MR) cnloh, default TRUE. if FALSE returns just the start/end}

\item{return.type}{(character) data.table or GRanges}

\item{general}{(logical) CNLOH in the general sense - where both alleles change in CN and not necessarily have to go to zero}

\item{balanced.only}{(logical) only balanced rearrangemetns without change in total CN}
}
\value{
LOH starting points plus metadata column "balanced" indicated whether it is copy-neutral
}
\description{
grab_cnloh
}
