% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{grab_jabba_epgaps}
\alias{grab_jabba_epgaps}
\title{grab_jabba_epgaps}
\usage{
grab_jabba_epgaps(pairs = NULL, mc.cores = 8, return.table = FALSE)
}
\arguments{
\item{pairs}{(keyed data.table)}

\item{mc.cores}{(numeric) default 8}

\item{return.table}{(logical) return data.table? default FALSE}
}
\value{
either named list of data.table
}
\description{
create named vector (or keyed table) of jabba epgaps
}
