% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{grab_loose}
\alias{grab_loose}
\title{grab_loose}
\usage{
grab_loose(
  jabba_rds,
  loose_ends = NA_character_,
  mask = "~/projects/gGnome/files/new.mask.rds",
  mask_pad = 10000,
  bad_thresh = 0.4,
  norm_thresh = 0.6,
  min_size = 50000,
  subset = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{jabba_rds}{(character) path to JaBbA}

\item{mask}{(character) path to coverage mask}

\item{mask_pad}{(numeric) pad on distance to mask}

\item{bad_thresh}{(numeric) default 0.3}

\item{norm_thresh}{(numeric) 0.6 (from filter.loose)}

\item{min_size}{(numeric) minimum size in BP}

\item{subset}{(logical) subset loose ends to include only valid ones before returning}

\item{verbose}{(logical)}
}
\description{
grab_loose
}
