% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{grab_loose_from_seg}
\alias{grab_loose_from_seg}
\title{grab_loose_from_seg}
\usage{
grab_loose_from_seg(
  gr = NULL,
  min.width = 0,
  max.gap = 1e+05,
  return.type = "data.table",
  verbose = FALSE
)
}
\arguments{
\item{gr}{(GRanges) with metadata field $cn}

\item{min.width}{(numeric) minimum SV width (default 0?)}

\item{max.gap}{(numeric) maximum gap between non-NA segments}

\item{verbose}{(logical)}
}
\value{
GRanges each with width 1 giving CN change points
}
\description{
get the location and direction of copy number increases
e.g. if the neighboring node on the left is lower, then strand is +
if the neighboring node on the right is lower, then strand is -
(opposite direction from paper, but consistent with loose ends from JaBbA)
}
