% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{grab_segs}
\alias{grab_segs}
\title{grab_segs}
\usage{
grab_segs(
  gg = "/dev/null",
  gr = "/dev/null",
  bw = "/dev/null",
  csv = "/dev/null",
  field = "cn",
  simplify = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{gg}{(character) path to jabba_rds}

\item{gr}{(character) path to GRanges (.rds file)}

\item{bw}{(character) path to bigWig (.bw)}

\item{csv}{(character) path to delimited file coercible to GRanges}

\item{simplify}{(logical) default TRUE. Merge adjacent ranges with the same copy number?}

\item{verbose}{(logical)}
}
\value{
GRanges with metadata field $cn
}
\description{
read CN segmentation from supplied gGraph, GRanges, bigWig, or delimited file coercible to GRanges
}
