% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{haplotype_support}
\alias{haplotype_support}
\title{haplotype_support}
\usage{
haplotype_support(wks, reads, min.consec = 3, return.bx = TRUE)
}
\arguments{
\item{wks}{(gWalk) haplotypes}

\item{reads}{(GRanges) putative supporting reads}
}
\value{
gWalk with n_supp_bx annotation giving the number of supporting barcodes per walk
}
\description{
haplotype_support
}
