% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{hets2roh}
\alias{hets2roh}
\title{hets2roh}
\usage{
hets2roh(
  sites,
  min.frac = 0.2,
  exclude.centromeres = FALSE,
  return.type = "GRanges"
)
}
\arguments{
\item{min.frac}{(numeric) minimum normal allele frequency default 0.2}

\item{exclude.centromeres}{(logical) default FALSE}

\item{return.type}{(character) GRanges or data.table}

\item{hets}{(character) path to sites.txt}
}
\value{
GRanges of all ROH
}
\description{
from sites.txt get a GRanges with heterozygous sites
}
