% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{inputdata}
\alias{inputdata}
\alias{inputdata,Zi-method}
\alias{inputdata<-}
\alias{inputdata<-,Zi-method}
\title{Access and Set the inputdata}
\usage{
inputdata(x)

\S4method{inputdata}{Zi}(x)

inputdata(x) <- value

\S4method{inputdata}{Zi}(x) <- value
}
\arguments{
\item{x}{\code{\linkS4class{Zi}}-class object}

\item{value}{inputdata object}
}
\value{
inputdata
}
\description{
access the inputdata of an \code{\linkS4class{Zi}}-class object
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
inputdata(Zi)
}
