% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{jabba_alt}
\alias{jabba_alt}
\title{jabba_alt}
\usage{
jabba_alt(gg = NULL, min_size = 0)
}
\arguments{
\item{gg}{(JaBbA gGraph)}

\item{min_size}{(numeric) min indel size, default 0}
}
\value{
Junction object with ALT junctions
}
\description{
jabba_alt
}
