% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{junction_coverage}
\alias{junction_coverage}
\title{junction_coverage}
\usage{
junction_coverage(
  rafile = NULL,
  cov_rds = NULL,
  standard.only = TRUE,
  field = "foreground",
  pad = 50000,
  min.span = 10000,
  small.inv = TRUE,
  mask = "~/projects/gGnome/files/lowmap.rds",
  verbose = FALSE
)
}
\arguments{
\item{rafile}{(character)}

\item{cov_rds}{(character)}

\item{standard.only}{(logical) default TRUE}

\item{field}{(character) default foreground}

\item{pad}{(character) default 5e4}

\item{mask}{(character) path to coverage mask}

\item{verbose}{(logical)}
}
\value{
data.table with columns:
- bp.index (junction index)
- bp (either 1 or 2, indicating first or second breakpoint)
- breakpoint (gr.string of breakpoint)
- junction (grl.string of junction)
- fused (indicating if coverage is from fused or unfused side)
- cn (coverage)
}
\description{
get coverage around the breakpoints of each junction
}
