% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{junction_mappability}
\alias{junction_mappability}
\title{junction_mappability}
\usage{
junction_mappability(
  junctions,
  ref,
  pad = 101,
  build = "hg19",
  verbose = FALSE
)
}
\arguments{
\item{junctions}{(gGnome Junction object or GRangesList)}

\item{ref}{(BWA object) reference genome}

\item{pad}{(numeric) number of base pairs padding, default 101}

\item{verbose}{(logical) default FALSE}
}
\value{
Junctions object with numeric metadata fields bp1.mapq and bp2.mapq
}
\description{
Given a set of junctions and a specified pad
adds fields bp1.mapq and bp2.mapq
referring to the MAPQ of each breakend
}
