% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{junctions_analysis}
\alias{junctions_analysis}
\title{junctions_analysis}
\usage{
junctions_analysis(
  gg = NULL,
  gs = "~/projects/JaBbA_pipeline/db/hcc1954.sv.gs.rds",
  max.dist = 1000,
  method = "jabba",
  id = "sample",
  mask = "~/projects/gGnome/files/new.mask.rds",
  verbose = FALSE
)
}
\arguments{
\item{gg}{(character) path to JaBbA gGraph}

\item{gs}{(character) path to SV GRanges}

\item{max.dist}{(numeric) maximum distance to consider junction overlap, default 1kb}

\item{method}{(character) default 'jabba', others not implemented yet}

\item{id}{(character)}

\item{mask}{(character) path to coverage mask (junctions in these regions are not considered.)}

\item{verbose}{(logical) default FALSE}

\item{res}{(list with elements)
- junctions.jj: Junctions object
- junctions.res: data.table with columns pair, precision, recall}
}
\description{
junctions_analysis
}
