% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{log1p}
\alias{log1p}
\alias{log1p,Zi-method}
\title{log(1+x)}
\usage{
\S4method{log1p}{Zi}(x)
}
\arguments{
\item{x}{\code{\linkS4class{Zi}}-class object}
}
\value{
a \code{\linkS4class{Zi}}-class object where the log(1+x) values of
\code{inputcounts, deinflatedcounts and weights} are calculated.
}
\description{
Calculate log(1+x) of all 'matrix' objects of a 'Zi'-class
object, log calculates by default natural logarithms
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
log1p(Zi)
}
\seealso{
\link[base]{log1p}, \link[zitools]{log2p}
}
