% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{log2p}
\alias{log2p}
\alias{log2p,Zi-method}
\title{log2p(x+1)}
\usage{
log2p(x)
}
\arguments{
\item{x}{\code{\linkS4class{Zi}}-class object}
}
\description{
Calculate log2(x+1) of all 'matrix' objects of a 'Zi'-class
object
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
log2p(Zi)

}
