% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{mask_gg_cn}
\alias{mask_gg_cn}
\title{mask_gg_cn}
\usage{
mask_gg_cn(
  jabba_rds,
  mask = "~/projects/gGnome/files/lowmap.rds",
  frac_thresh = 0.5,
  base_thresh = 1e+06,
  return.type = "GRanges",
  verbose = FALSE
)
}
\arguments{
\item{jabba_rds}{(character) path to jabba rds}

\item{mask}{(character) path to mask}

\item{base_thresh}{(numeric) min number of CN-supporting bases}

\item{return.type}{(character)}

\item{mask_thresh}{(numeric) between zero and one, fraction of width masked}
}
\description{
returns only the segments with less than mask_thresh overlapping the mask
}
