% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{mean}
\alias{mean}
\alias{mean.Zi}
\title{Arithmetic Mean}
\usage{
\method{mean}{Zi}(zi_result, ...)
}
\arguments{
\item{...}{}

\item{x}{An Object of class "Zi", input matrix will be used to calculate the
mean taking structural zero weights into account}
}
\description{
calculate the arithmetic mean of zero inflated data taking weights
for structural zeros into account
}
