% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{median}
\alias{median}
\alias{median.Zi}
\title{Calculate the median of zero-deinflated count data}
\usage{
\method{median}{Zi}(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{\code{\linkS4class{Zi}}-class object}

\item{na.rm}{logical, default = TRUE, NAs are excluded}

\item{...}{see \link[stats]{median.default}}
}
\value{
median value
}
\description{
Caluclate the median of  zero-deinflated data of a 'Zi'-class
object. To calculate the median, the output matrix will be extracted
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
median(Zi)
}
