% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{MissingValueHeatmap}
\alias{MissingValueHeatmap}
\title{Missing Value Heatmap}
\usage{
MissingValueHeatmap(ZiObject, title = "")

MissingValueHeatmap(ZiObject, title = "")
}
\arguments{
\item{ZiObject}{ZiObject, result of the ziMain function}
}
\value{
heatmap

heatmap
}
\description{
Missing Value Heatmap

Missing Value Heatmap
}
