% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{mnorm_cov}
\alias{mnorm_cov}
\title{mnorm_cov}
\usage{
mnorm_cov(
  cov_rds = "/dev/null",
  cov.gr = GRanges(),
  template.gr = GRanges(),
  template = "~/projects/dryclean/MPON_raw_inputs/WCM-4_1kb.rds",
  field = "auto",
  clean = TRUE,
  chrsub = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{cov_rds}{(character) path to coverage file}

\item{template.gr}{(GRanges) template as GRanges}

\item{template}{(character) path to template}

\item{field}{(character) default counts}

\item{chrsub}{(logical) get rid of chr prefix}

\item{verbose}{(logical)}
}
\value{
GRanges of cov_rds with counts coerced to template and mean normalized in $reads.corrected
}
\description{
mean normalize reads and coerce to template for dryclean
}
