% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtx}
\alias{mtx}
\title{Matrix Data}
\format{
\subsection{'mtx'}{

A matrix with 100 rows and 10 columns
}
}
\usage{
mtx
}
\value{
a data matrix
}
\description{
Zero-inflated matrix data
}
\keyword{datasets}
