% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{omit_str_zero}
\alias{omit_str_zero}
\title{Draw Structural Zeros}
\usage{
omit_str_zero(zi, zi_input, feature = "")
}
\arguments{
\item{zi}{result of fitting a zero inflation model to the data using pscl::zeroinfl}

\item{zi_input}{result of reshape_zi, count data in a long format}

\item{feature}{character string characterizing the rows, e.g. gene, OTU, ...}
}
\value{
a dataframe(long format), columns: count, sample, "feature",
}
\description{
draw structural zeros with a given probability (p_str_zero),
compare if predicted zero = zero count, if YES set to NA, repeat until
NA >= sum(p_str_zero), replace excess of NA with zero (random sampling)
}
