% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{otu_table}
\alias{otu_table}
\title{Access the otu table}
\usage{
\S4method{otu_table}{Zi}(object)
}
\arguments{
\item{ZiObject}{'Zi'-class object}
}
\value{
otu_table
}
\description{
access the otu_table of an 'Zi'-class object if the
datafile slot is a phyloseq object
}
