% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{overlap_breakpoints_with_jabba}
\alias{overlap_breakpoints_with_jabba}
\title{overlap_breakpoints_with_jabba}
\usage{
overlap_breakpoints_with_jabba(
  gr = GRanges(),
  jabba_rds = NA_character_,
  mask = "~/projects/gGnome/files/new.mask.rds",
  min.footprint = 0,
  bp.pad = 10000,
  mask.pad = 10000,
  verbose = FALSE
)
}
\arguments{
\item{gr}{(GRanges) of loose ends}

\item{jabba_rds}{(character) path to JaBbA output}

\item{mask}{(character) path to coverage mask}

\item{min.footprint}{(numeric) min SV size (default 1e4)}

\item{bp.pad}{(numeric) distance in bp from junction/loose end/mask (default 1e4)}

\item{mask.pad}{(numeric) distance in bp from junction/loose end/mask (default 1e4)}

\item{verbose}{(logical)}
}
\value{
GRanges with same length as input with three metadata columns:
$junction (logical)
$loose (logical)
$mask (logical)
}
\description{
Overlap a GRanges of loose ends with
the breakpoints in a JaBbA graph

Determine whether each loose end overlaps a junction, a loose end,
or overlaps an (optional) coverage mask.
}
