% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{phy_tree}
\alias{phy_tree}
\alias{phy_tree,Zi-method}
\title{Access the phylogenetic tree}
\usage{
\S4method{phy_tree}{Zi}(physeq, errorIfNULL = TRUE)
}
\arguments{
\item{physeq}{\code{\linkS4class{Zi}}-class object}

\item{errorIfNULL}{Logical. Should the accessor stop with an error if the slot
is empty (NULL)? Default TRUE.}
}
\value{
phy_tree
}
\description{
access the \link[phyloseq]{phy_tree} of an
\code{\linkS4class{Zi}}-class object if the inputdata slot is a phyloseq
object
}
