% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot}
\alias{plot}
\alias{plot,Zi-method}
\title{Plotting}
\usage{
\S4method{plot}{Zi,ANY}(x, y, ...)
}
\arguments{
\item{object}{\code{\linkS4class{Zi}}-class object}
}
\value{
returns plot object
}
\description{
plot
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
plot(Zi)

}
