% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{plot_pp}
\alias{plot_pp}
\title{plot_pp}
\usage{
plot_pp(
  tile.dt,
  min.cn = -0.5,
  max.cn = 7.5,
  bins = 400,
  ytrans = TRUE,
  vlines = TRUE,
  xlab = "Estimated CN",
  ylab = "Count"
)
}
\description{
create purity ploidy plot from tile output
}
