% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{+}
\alias{+}
\alias{+,Zi-method}
\title{Arithmetic Operators}
\usage{
\S4method{+}{Zi,ANY}(e1, e2)
}
\arguments{
\item{e1}{\code{\linkS4class{Zi}}-class object, matrix or number}

\item{e2}{\code{\linkS4class{Zi}}-class object, matrix or number}
}
\description{
Arithmetic operators for a Zi-class object
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
Zi+Zi
Zi+2

}
