% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{pp_tile}
\alias{pp_tile}
\title{pp_tile}
\usage{
pp_tile(
  seg = NA_character_,
  cov = NA_character_,
  kag = NA_character_,
  jab = NA_character_,
  purity = NA,
  ploidy = NA,
  min.quantile = 0.001,
  min.val = 0,
  max.quantile = 0.999,
  max.val = 100,
  field = "foreground",
  mask = "~/projects/gGnome/files/new.mask.rds",
  return.type = "data.table",
  tile.width = 1e+05
)
}
\arguments{
\item{seg}{(character) path to input segmentation}

\item{cov}{(character) path to coverage file}

\item{kag}{(character) path to karyograph}

\item{jab}{(character) path to JaBbA output}

\item{purity}{(numeric) overrides pp in jab}

\item{ploidy}{(numeric) overrides pp in jab}

\item{field}{(character) default "foreground" (signal name in cov)}

\item{mask}{(character) default ~/projects/gGnome/files/new.mask.rds}

\item{tile.width}{(numeric) default 1e5}
}
\value{
data.table of tiled genome with mean
}
\description{
tiles an input JaBbA graph or karyograph into bins of size tile.width

Creates a data.table representing tiled genome with column $cn
This can be used as input to pp_plot to QC purity/ploidy estimates
}
