% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocess_mtx}
\alias{preprocess_mtx}
\title{Preprocess Matrix}
\usage{
preprocess_mtx(mtx)
}
\arguments{
\item{mtx}{count matrix}
}
\value{
randomly sorted matrix
}
\description{
Preprocess matrix before subsetting it into blocks. The matrix is
randomly sorted to make sure that no prior sorting interferes when fitting the
model
}
