% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{proccov}
\alias{proccov}
\title{proccov}
\usage{
proccov(
  p,
  field = "tumor_cov_raw",
  proc.field = "^count$|^reads$",
  save = TRUE,
  template,
  mock = FALSE,
  force = FALSE,
  pairs
)
}
\description{
proccov
}
