% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{quantile}
\alias{quantile}
\alias{quantile.Zi}
\title{Calculate the quantiles of zero-deinflated count data}
\usage{
\method{quantile}{Zi}(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{A 'Zi'-class object}

\item{na.rm}{logical, default = TRUE, NAs are excluded}

\item{...}{\link[stats]{quantile}}
}
\value{
quantile values
}
\description{
Caluclate the quantiles of  zero-deinflated data of a 'Zi'-class
object. To calculate the quantiles, the output matrix will be extracted.
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
quantile(Zi)
}
