% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{remove_centromeres}
\alias{remove_centromeres}
\title{remove_centromeres}
\usage{
remove_centromeres(segs, pad = 0)
}
\arguments{
\item{segs}{(GRanges)}
}
\value{
input GRanges with ranges overlapping centromeres removed.
}
\description{
Remove segments in an input GRanges overlapping centromeres, preserving metadata.
}
