% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{remove_small_junctions}
\alias{remove_small_junctions}
\title{remove_small_junctions}
\usage{
remove_small_junctions(jj = jJ(), thresh = 10000, keep.inv = TRUE)
}
\arguments{
\item{jj}{(Junction) Junction object}

\item{thresh}{(numeric) min size (default 1e4)}

\item{keep.inv}{(logical) keep inversions? default TRUE}
}
\value{
Junction object
}
\description{
remove_small_junctions
}
