% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{replace_phyloseq}
\alias{replace_phyloseq}
\title{Replace the otu table of a phyloseq object}
\usage{
replace_phyloseq(ZiObject)
}
\arguments{
\item{ZiObject}{'Zi'-class object with a phyloseq object as input}
}
\value{
a "phyloseq"-class object
}
\description{
Replace the OTU table of a phyloseq object with an OTU table
where predicted structural zeros are replaced with NA
}
