% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{resample_output}
\alias{resample_output}
\title{Resample a \code{\linkS4class{Zi}}-class object}
\usage{
resample_output(x)
}
\arguments{
\item{x}{\code{\linkS4class{Zi}}-class object}
}
\description{
Resample the output matrix of an \code{\linkS4class{Zi}}-class object.
Resampling is done by drawing from a binomial distribution with a given probability
that a count value (zero and non-zero) is a structural zero.
}
