% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reshape_zi}
\alias{reshape_zi}
\title{Reshape a given matrix into long format}
\usage{
reshape_zi(mtx, feature = "")
}
\arguments{
\item{mtx}{count matrix}

\item{feature}{character string characterizing the rows, e.g. gene, OTU}
}
\value{
a dataframe (long format), 3 columns: count, sample, "feature"
}
\description{
reshape a given matrix into a long format as input for for fitting
a zero inflated model
}
