% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{rowData}
\alias{rowData}
\alias{rowData,Zi-method}
\title{Access the row data}
\usage{
\S4method{rowData}{Zi}(x, useNames = TRUE, ...)
}
\arguments{
\item{x}{\code{\linkS4class{Zi}}-class object}

\item{useNames}{returns a rowData dataframe with rownames}

\item{...}{\code{\link[SummarizedExperiment]{rowData}}}
}
\value{
DFrame
}
\description{
access the \link[SummarizedExperiment]{rowData} of an \code{\linkS4class{Zi}}-class object if the inputdata
is an object of the class SummarizedExperiment
}
