% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{rowQuantiles}
\alias{rowQuantiles}
\alias{colQuantiles}
\title{Calculates quantiles for each row (column) of a matrix-like object}
\usage{
\S4method{rowQuantiles}{Zi}(
  x,
  rows = NULL,
  cols = NULL,
  probs = seq(from = 0, to = 1, by = 0.25),
  na.rm = TRUE,
  type = 7L,
  ...,
  useNames = NA,
  drop = TRUE
)

\S4method{colQuantiles}{Zi}(
  x,
  rows = NULL,
  cols = NULL,
  probs = seq(from = 0, to = 1, by = 0.25),
  na.rm = TRUE,
  type = 7L,
  ...,
  useNames = NA,
  drop = TRUE
)
}
\arguments{
\item{x}{An Object of class "Zi", NA matrix will be extracted using this function}

\item{rows, }{A vector indicating the subset of rows (and/or columns) to operate}

\item{cols}{over. If NULL, no subsetting is done}

\item{probs}{A numeric vector of J probabilities in \link{0,1}}

\item{na.rm}{logical, default = TRUE, NAs are excluded}

\item{type}{An integer specifying the type of estimator}

\item{...}{Additional arguments passed to specific methods}

\item{useNames}{If NA, the default behavior of the function about naming support
is remained. If FALSE, no naming support is done. Else if TRUE, names
attributes of result are set.}

\item{drop}{If TRUE a vector is returned if J == 1.}
}
\description{
Calculates quantiles for each row (column) of a matrix-like object
}
