% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{rowSds}
\alias{rowSds}
\title{Row and Column Standard Deviations}
\usage{
\S4method{rowSds}{Zi}(x)
}
\arguments{
\item{x}{An Object of class "Zi", input matrix will be used to calculate the
row or col Standard Deviation taking structural zero weights into account}
}
\value{
a vector of row/col length
}
\description{
calculate row and column sds for matrix like objects
}
