% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{rowWeightedSds}
\alias{rowWeightedSds}
\alias{rowWeightedSds,Zi-method}
\alias{colWeightedSds}
\alias{colWeightedSds,Zi-method}
\alias{rowWeightedVars}
\alias{rowWeightedVars,Zi-method}
\alias{colWeightedVars}
\alias{colWeightedVars,Zi-method}
\title{Row and column weighted standard deviations or variances of zero
inflated count data}
\usage{
\S4method{rowWeightedSds}{Zi}(x, w, rows = NULL, cols = NULL, na.rm = FALSE, useNames = TRUE)

\S4method{colWeightedSds}{Zi}(x, w, rows = NULL, cols = NULL, na.rm = FALSE, useNames = TRUE)

\S4method{rowWeightedVars}{Zi}(x, w, rows = NULL, cols = NULL, na.rm = FALSE, useNames = TRUE)

\S4method{colWeightedVars}{Zi}(x, w, rows = NULL, cols = NULL, na.rm = FALSE, useNames = TRUE)
}
\arguments{
\item{x}{A \code{\linkS4class{Zi}}-class object}

\item{w}{a numerical vector of weights either of length = rows or length = cols
giving the weights to use for elements of x}

\item{rows, cols}{A \code{\link[base]{vector}} indicating the subset of rows and/or columns to
operate over. If \code{\link{NULL}} (default), no subsetting is done}

\item{na.rm}{\code{\link[base]{logical}} If \code{\link{TRUE}} \code{\link{NA}}s are excluded, otherwise
not. default = \code{\link{FALSE}}}

\item{useNames}{\code{\link[base]{logical}}  If \code{\link{TRUE}} (default), names
attributes of result are set. Else if \code{\link{FALSE}}, no naming support is done.}
}
\value{
a numeric vector of length N(K)
}
\description{
Calculate row and column standard deviations or variances of
zero inflated count data, additionally taking weights for structural zeros
into account.
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
rowWeightedSds(Zi, w = runif(ncol(Zi@countmatrix), 0.1,1))
colWeightedSds(Zi, w = runif(nrow(Zi@countmatrix), 0.1,1))
rowWeightedVars(Zi, w = runif(ncol(Zi@countmatrix), 0.1,1))
colWeightedVars(Zi, w = runif(nrow(Zi@countmatrix), 0.1,1))

}
