% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{rowWeightedVars}
\alias{rowWeightedVars}
\alias{colWeightedVars}
\title{Calculates the weighted variance for each row (column) of a matrix-like object}
\usage{
\S4method{rowWeightedVars}{Zi}(
  x,
  w = NULL,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  ...,
  useNames = NA
)

\S4method{colWeightedVars}{Zi}(
  x,
  w = NULL,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  ...,
  useNames = NA
)
}
\arguments{
\item{x}{An Object of class "Zi", input matrix will be used to calculate the
row weighted variances taking structural zero weights into account}

\item{w}{a numerical vector of weight the same length as x giving the weights
to use for elements of x}

\item{...}{}
}
\value{
a numeric vector of length N(K)
}
\description{
Calculates the row weighted variances for each row (column) of a matrix-like object.
}
\examples{
rowWeightedVars(Zi, w = runif(ncol(Zi@countmatrix), 0.1,1))
colWeightedVars(Zi, w = runif(nrow(Zi@countmatrix), 0.1,1)

}
