% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{sample_granges}
\alias{sample_granges}
\title{sample_granges}
\usage{
sample_granges(
  gr,
  k,
  width = 1,
  verbose = FALSE,
  seqlengths = NULL,
  mc.cores = 8
)
}
\arguments{
\item{gr}{(GRanges)}

\item{k}{(numeric) number of samples}

\item{width}{(numeric) width of each sample}
}
\description{
sample_granges
}
