% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{sd}
\alias{sd}
\title{Standard Deviation of zero inflated count data}
\usage{
\S4method{sd}{Zi}(x)
}
\arguments{
\item{x}{A 'Zi'-class object}

\item{...}{\link[stats]{sd} or \link[matrixStats]{weightedSd}}
}
\value{
value
}
\description{
Calculate the standard deviation of zero inflated count data
taking weights for structural zeros into account.
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
sd(Zi)
}
