% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{show}
\alias{show}
\alias{show,Zi-method}
\title{Show summary of Zi object}
\usage{
\S4method{show}{Zi}(object)
}
\arguments{
\item{object}{\code{\linkS4class{Zi}}-class object}
}
\value{
returns a numeric vector of row/column length
}
\description{
Message printed at command line
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
Zi
show(Zi)

}
