% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{standard_grl}
\alias{standard_grl}
\title{standard_grl}
\usage{
standard_grl(grl = GRangesList())
}
\arguments{
\item{grl}{}
}
\value{
grl but filtered to only include standard chromosomes
}
\description{
standard_grl
}
