% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{starcode}
\alias{starcode}
\title{starcode}
\usage{
starcode(x)
}
\arguments{
\item{x}{(numeric) vector of p values}
}
\value{
character vector of stars corresponding with p value magnitude
}
\description{
convert p values to stars
}
