% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{subset_feature}
\alias{subset_feature}
\title{Subset a \code{\linkS4class{Zi}}-class object based on feature data}
\usage{
subset_feature(Zi, ...)
}
\arguments{
\item{Zi}{\code{\linkS4class{Zi}}-class object}

\item{...}{The subsetting expression that should be applied, see \link[base]{subset}
for more details}
}
\description{
Subset a \code{\linkS4class{Zi}}-class object based on tax_table
of a phyloseq
object or on rowData of a SummarizedExperiment object
}
