% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_mtx}
\alias{subset_mtx}
\title{subset matrix into groups of roughly 5000 datapoints per block}
\usage{
subset_mtx(mtx)
}
\arguments{
\item{mtx}{matrix to be subsetted}
}
\value{
list containing the subsets
}
\description{
subset a matrix into blocks of approximately 5000 datapoints per
block.
}
