% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{subset_sample}
\alias{subset_sample}
\alias{subset_feature}
\title{Subset a 'Zi'-class object based on sample data}
\usage{
subset_sample(Zi, ...)

subset_feature(Zi, ...)

subset_feature(Zi, ...)
}
\arguments{
\item{Zi}{'Zi'-class object}

\item{...}{The subsetting expression that should be applied, see \link[base]{subset}
for more details}
}
\description{
Subset a 'Zi'-class object based on sample_data of an phyloseq
object or on colData based on a SummarizedExperiment object

Subset a 'Zi'-class object based on sample_data of an phyloseq
object or on colData of a SummarizedExperiment object

Subset a 'Zi'-class object based on tax_table of a phyloseq
object or on rowData of a SummarizedExperiment object
}
