% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{svbnd}
\alias{svbnd}
\title{svbnd}
\usage{
svbnd(svs, field = "SVTYPE", return.type = "grl")
}
\arguments{
\item{svs}{(GRanges) with annotation field SVTYPE (or character vec specified in field)}

\item{field}{(character) default SVTYPE}

\item{return.type}{(character) one of jJ, grl, default jJ}
}
\description{
Helper function I stole from Marcin's blog file for converting ranges to breakends
}
