% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{t}
\alias{t}
\title{Transpose a 'Zi'-class object}
\usage{
\S4method{t}{Zi}(x)
}
\arguments{
\item{ZiObject}{'Zi'-class object}
}
\value{
'Zi'-class object
}
\description{
transpose all matrizes of a 'Zi'-class object
}
