% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{telomeric_junctions}
\alias{telomeric_junctions}
\title{telomeric_junctions}
\usage{
telomeric_junctions(junctions, pad = 101, verbose = FALSE)
}
\arguments{
\item{junctions}{(gGnome Junction object or GRangesList)}

\item{pad}{(numeric) number of base pairs padding, default 101}

\item{verbose}{(logical) default FALSE}
}
\value{
Junctions object with logical metadata fields bp1.telomere and bp2.telomere
}
\description{
Given a set of junctions
identify whether bp1/bp2 contains a telomere trimerf
on the (+) strand adjacent to the junction breakend
}
