% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{valid_walks}
\alias{valid_walks}
\title{valid_walks}
\usage{
valid_walks(junctions, pad = 20000, max_dist = 1e+06)
}
\arguments{
\item{junctions}{(gGnome::Junction)}

\item{pad}{(numeric) how much to pad the sides of walks (default 2e4)}

\item{max_dist}{(numeric) default 1 Mbp}
}
\value{
gWalks
}
\description{
valid_walks
}
