% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{var}
\alias{var}
\title{Variance}
\usage{
\S4method{var}{Zi}(x)
}
\arguments{
\item{x}{An Object of class "Zi", input matrix will be used to calculate the
variance taking structural zero weights into account}

\item{...}{}
}
\value{
value
}
\description{
calculate the variance of zero inflated data taking weights
for structural zeros into account
}
\examples{
var(Zi)
}
