% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{weighted.mean}
\alias{weighted.mean}
\title{Weighted Arithmetic Mean}
\usage{
\S4method{weighted.mean}{Zi}(x, w, ...)
}
\arguments{
\item{x}{An Object of class "Zi", input matrix will be used to calculate the
meam taking structural zero weights into account}

\item{w}{a numerical vector of weight the same length as x giving the weights
to use for elements of x}

\item{...}{}
}
\value{
value
}
\description{
compute a weighted mean
}
\examples{
weight <- runif(length(Zi@countmatrix), 0.1, 1)
weighted.mean(Zi, w= weight)
}
