% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{weightedVar}
\alias{weightedVar}
\title{Weighted Variance and weighted Standard Deviation}
\usage{
weightedVar(x, w = NULL, idxs = NULL, na.rm = FALSE, center = NULL, ...)
}
\arguments{
\item{x}{An Object of class "Zi", input matrix will be used to calculate the
row weighted Standard Deviation taking structural zero weights into account}

\item{w}{a vector of weights the same length as x giving the weights to use for
each element of x. Negative weights are treated as zero weights. Default value
is equal weight to all values.}
}
\value{
a numeric scalar
}
\description{
Computes a weighted variance / standard deviation of a numeric
vector or across rows or columns of a matrix
}
