% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{weights}
\alias{weights}
\alias{weights,Zi-method}
\title{Access the weights}
\usage{
weights(x)

\S4method{weights}{Zi}(x)
}
\arguments{
\item{x}{\code{\linkS4class{Zi}}-class object}

\item{value}{replacement of deinflatedcounts}
}
\value{
weights
}
\description{
access the weights of an \code{\linkS4class{Zi}}-class object
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
weights(Zi)
}
