% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{zi2deinflatedcounts}
\alias{zi2deinflatedcounts}
\title{Access the zero de-inflated matrix of an \code{\linkS4class{Zi}}-class object}
\usage{
zi2deinflatedcounts(ZiObject)
}
\arguments{
\item{ZiObject}{\code{\linkS4class{Zi}}-class object}
}
\value{
matrix
}
\description{
extract the zero de-inflated matrix of an \code{\linkS4class{Zi}}-class object. The
deinflatedcounts matrix is a count matrix (column = sample, row = feature) where
predicted structural zeros are replaced with NA
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
zi2deinflatedcounts(Zi)

}
