% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{zi2inputcounts}
\alias{zi2inputcounts}
\title{Access the count matrix of an \code{\linkS4class{Zi}}-class object}
\usage{
zi2inputcounts(ZiObject)
}
\arguments{
\item{ZiObject}{\code{\linkS4class{Zi}}-class object}
}
\value{
matrix
}
\description{
extract the count matrix (column = sample, row = feature), that
has been used to fit a zero inflation model
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
zi2inputcounts(Zi)
}
