% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{zi2inputMatrix}
\alias{zi2inputMatrix}
\alias{zi2countmatrix}
\title{Access the input matrix of an 'Zi'-class object}
\usage{
zi2countmatrix(ZiObject)
}
\arguments{
\item{ZiObject}{'Zi'-class object}
}
\value{
matrix
}
\description{
extract the count matrix (column = sample, row = feature), that
has been used to fit a zero inflation model
}
