% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{zi2outputMatrix}
\alias{zi2outputMatrix}
\title{Access the output matrix of an 'Zi'-class object}
\usage{
zi2outputMatrix(ZiObject)
}
\arguments{
\item{ZiObject}{'Zi'-class object}
}
\value{
matrix
}
\description{
extract the output matrix of an 'Zi'-class object. The
output matrix is a count matrix (column = sample, row = feature) where
predicted structural zeros are replaced with NA
}
