% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{zi2phyloseq}
\alias{zi2phyloseq}
\title{Replace the otu table of a phyloseq object}
\usage{
zi2phyloseq(ZiObject)
}
\arguments{
\item{ZiObject}{\code{\linkS4class{Zi}}-class object with a phyloseq object as input}
}
\value{
a "phyloseq"-class object
}
\description{
Replace the OTU table of a phyloseq object with the OTU table
of zero de-inflated count data
}
