% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{zi_to_deseq2}
\alias{zi_to_deseq2}
\title{Convert zi data to DESeq2 dds object}
\arguments{
\item{zi}{Zi-class object}

\item{design}{A formula which specifies the design of the experiment, taking
the form formula(~ x + y + z). That is, a formula with right-hand side only.
By default, the functions in this package and DESeq2 will use the last variable
in the formula (e.g. z) for presenting results (fold changes, etc.) and plotting.
When considering your specification of experimental design, you will want to re-order the
levels so that the NULL set is first.}

\item{colData}{for matrix input: a DataFrame or data.frame with at least a single
column. Rows of colData correspond to columns of countData}
}
\description{
Zi data is converted to a DESeqDataSet object, which can be used for
DESeq2 Analysis. Both, weight and count matrices will be stored in assays of
the DESeqDataSet
}
