% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zi_function.R
\name{ziMain}
\alias{ziMain}
\title{ziMain}
\usage{
ziMain(
  datafile,
  feature = "",
  formula,
  dist = c("poisson", "negbin", "geometric"),
  link = c("logit", "probit", "cloglog", "cauchit", "log"),
  zeroRows.rm = FALSE,
  ...
)
}
\arguments{
\item{datafile}{matrix (rows = features, columns = samples), phyloseq object or
SummarizedExperiment object}

\item{feature}{"gene", "OTU", "phylum", etc.}

\item{formula}{formula to fit the model response ~ predictor1 + predictor2 + ..., default = count ~ sample+OTU}

\item{dist}{= distribution, either poisson, negative binomial = negbin or geometric}

\item{link}{= link function, either logit, probit, cloglog, cauchit}

\item{zeroRows.rm}{= logical, TRUE if rows that only contain zeros should be removed
or not (they are removed to fit a zero inflated model and will be added afterwards
count matrix per default = 0 and weights = 1)}
}
\value{
S4 list with slots for the input object, the extracted count matrix if
the input is not a matrix, the results of the fitted zero inflation model, a
matrix where the predicted structural zeros are replaced with NA, a matrix
containing the calculated weights
}
\description{
The ziMain function uses a matrix, phyloseq, or SummarizedExperiment
object, extracts the count matrix to fit a zero inflation model to the data.
The matrix is divided into blocks of around 5000 count values to improve run time.
Predicted probabilities given that a zero in the count matrix  is a
structural zero are used to draw structural zeros and replace them with NA. Further,
weights for all zero counts are calculated given the following formula:
w=...
}
