% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zitools-package.R
\docType{package}
\name{zitools-package}
\alias{zitools}
\alias{zitools-package}
\title{zitools: Analysis of zero-inflated count data}
\description{
zitools allows for zero inflated count data analysis by either using down-weighting of excess zeros or by replacing an appropriate proportion of excess zeros with NA. Through overloading frequently used statistical functions (such as mean, median, standard deviation), plotting functions (such as boxplots or heatmap) or differential abundance tests, it allows a wide range of downstream analyses for zero-inflated data in a less biased manner. This becomes applicable in the context of microbiome analyses, where the data is often overdispersed and zero-inflated, therefore making data analysis extremly challenging.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kreutz-lab/zitools}
  \item Report bugs at \url{https://github.com/kreutz-lab/zitools/issues}
}

}
\author{
\strong{Maintainer}: Carlotta Meyring \email{carlotta.meyring@uniklinik-freiburg.de} (\href{https://orcid.org/0009-0000-6201-7615}{ORCID})

}
\keyword{internal}
