% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{assays}
\alias{assays}
\alias{assays,Zi-method}
\title{Access assays}
\usage{
\S4method{assays}{Zi}(x, withDimnames = TRUE, ...)
}
\arguments{
\item{x}{\code{\linkS4class{Zi}}-class object}

\item{withDimnames}{A \code{logical}, indicating whether the dimnames of the
SummarizedExperiment object should be applied (i.e. copied) to the extracted
assays. see \code{\link[SummarizedExperiment]{assays}}}

\item{...}{see \code{\link[SummarizedExperiment]{assays}}}
}
\value{
list
}
\description{
access  assays of an \code{\linkS4class{Zi}}-class object if the inputdata is an
object of the class SummarizedExperiment
}
