% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{boxplot}
\alias{boxplot}
\alias{boxplot.Zi}
\title{Box Plots}
\usage{
boxplot.Zi(x, log1p = FALSE, ...)
}
\arguments{
\item{x}{ZiObject, result of the ziMain function}

\item{log1p}{logical, default = FALSE, if TRUE log(1+p) transformation takes
place}

\item{...}{see boxplot documentation}
}
\description{
Box Plots
}
