% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{cn_analysis}
\alias{cn_analysis}
\title{cn_analysis}
\usage{
cn_analysis(
  seg = NULL,
  gs = NULL,
  method = "jabba",
  id = "sample",
  tile.width = 10000,
  min.width = 1000,
  mask = "~/projects/gGnome/files/new.mask.rds",
  verbose = FALSE
)
}
\arguments{
\item{seg}{(character) path to seg file}

\item{gs}{(character) path to gold standard segs}

\item{method}{(character) analysis method (one of jabba, ascat)}

\item{id}{(character) sample ID}

\item{tile.width}{(numeric) resolution at which to tile the genome for correlation. Default 1e4.}

\item{min.width}{(numeric) min tile width to consider for RMSE, etc.}

\item{mask}{(character) path to coverage mask}

\item{verbose}{(logical) default FALSE}
}
\value{
list with entries $cncp.gr $cncp.res
}
\description{
This is a function that compares segment CNs to a gold standard.
}
