% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{colMeans}
\alias{colMeans}
\alias{rowMeans}
\title{Row and Column Means}
\usage{
\S4method{colMeans}{Zi}(x)

\S4method{rowMeans}{Zi}(x)
}
\arguments{
\item{x}{An Object of class "Zi", input matrix will be used to calculate the
mean taking structural zero weights into account}
}
\value{
a vector of row/col length
}
\description{
calculate row and column means for matrix like objects
}
