% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{colMedians}
\alias{colMedians}
\title{Calculates the median for each row (column) of a matrix-like object}
\usage{
\S4method{colMedians}{Zi}(x, rows = NULL, cols = NULL, na.rm = TRUE, ..., useNames = NA)
}
\arguments{
\item{x}{An Object of class "Zi", NA matrix will be extracted using this function}

\item{rows, }{A vector indicating the subset of rows (and/or columns) to operate}

\item{cols}{over. If NULL, no subsetting is done}

\item{na.rm}{logical, default = TRUE, NAs are excluded}

\item{...}{Additional arguments passed to specific methods}

\item{useNames}{If NA, the default behavior of the function about naming support
is remained. If FALSE, no naming support is done. Else if TRUE, names
attributes of result are set.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      returns a numeric vector of length N(K)
}\if{html}{\out{</div>}}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Calculates the median for each row (column) of a matrix-like object.
             Prior to calculating the median, the NA matrix is extracted
}\if{html}{\out{</div>}}
}
