% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{colMedians}
\alias{colMedians}
\alias{rowMedians}
\title{Calculate the row or column median of zero-deinflated count data}
\usage{
\S4method{colMedians}{Zi}(x, rows = NULL, cols = NULL, na.rm = TRUE, ..., useNames = NA)

\S4method{rowMedians}{Zi}(x, rows = NULL, cols = NULL, na.rm = TRUE, ..., useNames = NA)
}
\arguments{
\item{x}{'Zi'-class object}

\item{rows, }{A vector indicating the subset of rows (and/or columns) to operate}

\item{cols}{over. If NULL, no subsetting is done}

\item{na.rm}{logical, default = TRUE, NAs are excluded}

\item{...}{\link\link[MatrixGenerics:rowMedians]{MatrixGenerics::rowMedians}}

\item{useNames}{If NA, the default behavior of the function about naming support
is remained. If FALSE, no naming support is done. Else if TRUE, names
attributes of result are set.}
}
\value{
returns a numeric vector of row/column length
}
\description{
Calculate the row or column median of  zero-deinflated data of a
'Zi'-class object. To calculate the median, the output matrix will be extracted
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
colMedians(Zi, useNames = TRUE)
rowMedians(Zi, useNames = TRUE)

}
