% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cor}
\alias{cor}
\alias{cor,Zi,ANY-method}
\title{Calculate weighted Pearson Correlation coeffiecients}
\usage{
\S4method{cor}{Zi,ANY}(x, y = NULL, use = "everything", method = "pearson")
}
\arguments{
\item{x}{'Zi'-class object}

\item{y}{'Zi'-class object}

\item{use}{"everything" see \link[stats]{cor}}

\item{method}{default = "pearson", weighted correlation only implemented for
person correlation}
}
\description{
calculate the weighted pearson correlation coefficients of a
count matrix of an Zi object taking weights for zero counts into account
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
cor(Zi)
}
