% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{ecycles}
\alias{ecycles}
\title{ecycles}
\usage{
ecycles(junctions = NULL, jabba_rds = NULL, thresh = 1000, verbose = FALSE)
}
\arguments{
\item{junctions}{(character) path to junctions file or GRangesList}

\item{jabba_rds}{(character) path to JaBbA output}

\item{thresh}{(numeric) distance threshold for reciprocality, default 1e3 (bp)}

\item{verbose}{(logical) default FALSE}
}
\value{
data.table with columns cycle.id, edge.id, class, junction
}
\description{
Return a data table of all ALT edges in a gGraph that is part of a simple cycle
}
