% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{heatmap}
\alias{heatmap}
\title{Draw a Heat Map}
\value{
heatmap
}
\description{
draw a heatmap of a given ziObject, heatmap.Zi uses the output
matrix (drawn structural zeros) to produce a heatmap. NA values are white
}
