% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{inputdata}
\alias{inputdata}
\alias{inputdata,Zi-method}
\title{Access the inputdata}
\usage{
inputdata(x)

\S4method{inputdata}{Zi}(x)
}
\arguments{
\item{x}{\code{\linkS4class{Zi}}-class object}
}
\value{
inputdata
}
\description{
access the inputdata of an \code{\linkS4class{Zi}}-class object
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
inputdata(Zi)
}
