% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{log1p}
\alias{log1p}
\title{log(1+x)}
\usage{
\S4method{log1p}{Zi}(x)
}
\arguments{
\item{x}{'Zi'-class object}
}
\description{
log1p(x) computs log(1+x) of all 'matrix' objects of a 'Zi'-class
object
}
\seealso{
\link[base]{log1p}
}
