% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{mean}
\alias{mean}
\alias{mean.Zi}
\title{Arithmetic Mean}
\usage{
\method{mean}{Zi}(zi_result, ...)
}
\arguments{
\item{...}{\link[base]{mean}}

\item{x}{A \code{\linkS4class{Zi}}-class object}
}
\value{
value
}
\description{
Calculate the arithmetic mean of zero inflated data taking weights
for structural zeros into account
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
mean(Zi)

}
