% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{median}
\alias{median}
\alias{median.Zi}
\title{Median Value}
\usage{
\method{median}{Zi}(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{'Zi'-class object, output matrix will be extracted using this function}

\item{na.rm}{logical, default = TRUE, NAs are excluded}

\item{...}{see \link[stats]{median.default}}
}
\value{
median value
}
\description{
Computes the median of the output matrix of an 'Zi'-class object
}
\examples{
median(Zi)
}
