% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{median}
\alias{median}
\alias{median.Zi}
\title{Median Value}
\usage{
median.Zi(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{An Object of class "Zi", NA matrix will be extracted using this function}

\item{na.rm}{logical, default = TRUE, NAs are excluded}
}
\value{
median value
}
\description{
computes the median of the NA matrix of an S4 "Zi" class object
}
