% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{MissingValueHeatmap}
\alias{MissingValueHeatmap}
\title{Missing Value Heatmap}
\usage{
MissingValueHeatmap(ZiObject, title = "", xlab = "", ylab = "")
}
\arguments{
\item{ZiObject}{ZiObject, result of the ziMain function}

\item{title}{Title of the plot .}

\item{xlab}{Title of the x axis.}

\item{ylab}{Title of the y axis.}
}
\value{
heatmap
}
\description{
Missing Value Heatmap
}
