% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{model}
\alias{model}
\alias{model,Zi-method}
\title{Access the model}
\usage{
model(x)

\S4method{model}{Zi}(x)
}
\arguments{
\item{x}{\code{\linkS4class{Zi}}-class object}
}
\value{
model
}
\description{
access the model of an \code{\linkS4class{Zi}}-class object
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
model(Zi)
}
