% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{phy_tree}
\alias{phy_tree}
\alias{phy_tree,Zi-method}
\title{Access the phylogenetic tree}
\usage{
\S4method{phy_tree}{Zi}(physeq)
}
\arguments{
\item{physeq}{\code{\linkS4class{Zi}}-class object}
}
\value{
phy_tree
}
\description{
access the phylogenetic tree (\link[phyloseq]{phy_tree}) of an object of the class
"Zi" if the inputdata slot is a phyloseq object
}
