% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{phy_tree}
\alias{phy_tree}
\alias{phy_tree,Zi-method}
\title{Access the phylogenetic tree}
\usage{
\S4method{phy_tree}{Zi}(physeq, errorIfNULL = TRUE)
}
\arguments{
\item{physeq}{\code{\linkS4class{Zi}}-class object}

\item{errorIfNULL}{Logical. Should the accessor stop with an error if the slot
is empty (NULL)? Default TRUE.}
}
\value{
phy_tree
}
\description{
access the phylogenetic tree (\link[phyloseq]{phy_tree}) of an object of the class
"Zi" if the inputdata slot is a phyloseq object
}
\examples{
data(mtx)
OTU <- otu_table(mtx, taxa_are_rows = TRUE)
sample_data <- data.frame(SampleID = c("Sample1", "Sample2", "Sample3", "Sample4",
                     "Sample5", "Sample6", "Sample7", "Sample8", "Sample9",
                     "Sample10"),
                     Group = factor(x = c(1,1,1,1,1,2,2,2,2,2)))
SAM <- sample_data(sample_data)
tax_table <- data.frame(Kingdom = c(rep("Bacteria", times = 100)),
                     Phylum = c(rep("Bacteroidetes", times = 50), rep("Firmicutes", times = 50)))
TAX <- tax_table(tax_table)
PHY <-
ps <- phyloseq::phyloseq(OTU, TAX, SAM)
Zi <- ziMain(ps)
phy_tree(Zi, errorIfNULL = FALSE)

}
