% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot}
\alias{plot}
\alias{plot,Zi,ANY-method}
\title{Plotting}
\usage{
\S4method{plot}{Zi,ANY}(x, y, ...)
}
\arguments{
\item{x}{\code{\linkS4class{Zi}}-class object}

\item{y}{the y coordinates of points in the plot, optional if x is an
appropriate structure}

\item{...}{Arguments to be passed to plot}
}
\value{
returns plot object
}
\description{
plot
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
plot(Zi)

}
