% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{quantile}
\alias{quantile}
\alias{quantile.Zi}
\title{Sample Quantiles}
\usage{
\method{quantile}{Zi}(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{An S4 object of class "Zi", NA matrix will be extracted using this function}

\item{na.rm}{logical, default = TRUE, NAs are excluded}
}
\value{
quantiles
}
\description{
computes the quantiles of the NA matrix of an S4 "Zi" class object
}
\examples{
quantile(Zi)
}
