% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{rowData}
\alias{rowData}
\title{Access the row data}
\usage{
\S4method{rowData}{Zi}(x, use.names = TRUE, ...)
}
\arguments{
\item{ZiObject}{\code{\linkS4class{Zi}}-class object}
}
\value{
DFrame
}
\description{
access the rowData of an \code{\linkS4class{Zi}}-class object if the inputdata
is an object of the class SummarizedExperiment
}
