% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{rowVars}
\alias{rowVars}
\alias{colVars}
\title{Row and Column Variances of zero inflated count data}
\usage{
\S4method{rowVars}{Zi}(x)

\S4method{colVars}{Zi}(x)
}
\arguments{
\item{x}{A 'Zi'-class object}
}
\value{
a vector of row/col length
}
\description{
Calculate row and column variances of zero inflated count data
taking weights for structural zeros into account.
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
rowVars(Zi)
colVars(Zi)
}
