% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{rowVars}
\alias{rowVars}
\title{Row and Column Variances}
\usage{
\S4method{rowVars}{Zi}(x)
}
\arguments{
\item{x}{An Object of class "Zi", input matrix will be used to calculate the
row or col Variance  taking structural zero weights into account}
}
\value{
a vector of row/col length
}
\description{
calculate row and column variances  for matrix like objects
}
