% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{rowWeightedMeans}
\alias{rowWeightedMeans}
\alias{colWeightedMeans}
\title{Row and Column weighted means of zero inflated count data}
\usage{
\S4method{rowWeightedMeans}{Zi}(
  x,
  w = NULL,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  ...,
  useNames = NA
)

\S4method{colWeightedMeans}{Zi}(
  x,
  w = NULL,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  ...,
  useNames = NA
)
}
\arguments{
\item{x}{A 'Zi'-class object}

\item{w}{a numerical vector of weight either of length = rows or length = cols
giving the weights to use for elements of x}

\item{...}{\link[MatrixGenerics]{rowWeightedMeans}}
}
\value{
a numeric vector of length N(K)
}
\description{
Calculate the weighted mean for each row (column) of a matrix-like object.
}
\examples{
rowWeightedMeans(Zi, w = runif(ncol(Zi@countmatrix), 0.1,1))
colWeightedMeans(Zi, w = runif(nrow(Zi@countmatrix), 0.1,1)

}
