% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{rowWeightedSds}
\alias{rowWeightedSds}
\alias{colWeightedSds}
\alias{rowWeightedVars}
\alias{colWeightedVars}
\title{Row and column weighted standard deviations or variances of zero
inflated count data}
\usage{
\S4method{rowWeightedSds}{Zi}(
  x,
  w = NULL,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  ...,
  useNames = NA
)

\S4method{colWeightedSds}{Zi}(
  x,
  w = NULL,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  ...,
  useNames = NA
)

\S4method{rowWeightedVars}{Zi}(
  x,
  w = NULL,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  ...,
  useNames = NA
)

\S4method{colWeightedVars}{Zi}(
  x,
  w = NULL,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  ...,
  useNames = NA
)
}
\arguments{
\item{x}{A "Zi"-class object}

\item{w}{a numerical vector of weights either of length = rows or length = cols
giving the weights to use for elements of x}

\item{...}{\link[matrixStats]{weightedSd}}
}
\value{
a numeric vector of length N(K)
}
\description{
Calculate row and column standard deviations or variances of
zero inflated count data, additionally taking weights for structural zeros
into account.
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
rowWeightedSds(Zi, w = runif(ncol(Zi@countmatrix), 0.1,1))
colWeightedSds(Zi, w = runif(nrow(Zi@countmatrix), 0.1,1))
rowWeightedVars(Zi, w = runif(ncol(Zi@countmatrix), 0.1,1))
colWeightedVars(Zi, w = runif(nrow(Zi@countmatrix), 0.1,1)

}
