% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{rowWeightedSds}
\alias{rowWeightedSds}
\title{Calculates the weighted mean for each row (column) of a matrix-like object}
\usage{
\S4method{rowWeightedSds}{Zi}(
  x,
  w = NULL,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  ...,
  useNames = NA
)
}
\arguments{
\item{x}{An Object of class "Zi", input matrix will be used to calculate the
row weighted Standard Deviation taking structural zero weights into account}

\item{w}{a numerical vector of weight the same length as x giving the weights
to use for elements of x}

\item{...}{}
}
\value{
a numeric vector of length N(K)
}
\description{
Calculates the row weighted mean for each row (column) of a matrix-like object.
}
