% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{sample_data}
\alias{sample_data}
\title{Access the sample data}
\usage{
\S4method{sample_data}{Zi}(object)
}
\arguments{
\item{ZiObject}{'Zi'-class object}
}
\value{
sample_data
}
\description{
access the sample_data of an 'Zi'-class object if the
inputdata slot is a phyloseq object
}
