% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{sd}
\alias{sd}
\title{Standard Deviation}
\usage{
\S4method{sd}{Zi}(x)
}
\arguments{
\item{x}{An Object of class "Zi", input matrix will be used to calculate the
standard deviation taking structural zero weights into account}

\item{...}{}
}
\value{
value
}
\description{
calculate the standard deviation of zero inflated data taking weights
for structural zeros into account
}
\examples{
sd(Zi)
}
