% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{sd}
\alias{sd}
\alias{sd,Zi-method}
\title{Standard Deviation of zero inflated count data}
\usage{
\S4method{sd}{Zi}(x, na.rm = FALSE)
}
\arguments{
\item{x}{A \code{\linkS4class{Zi}}-class object}

\item{na.rm}{\code{\link[base]{logical}} If \code{\link{TRUE}} \code{\link{NA}}s
are excluded, otherwise not. default = \code{\link{FALSE}}}
}
\value{
value
}
\description{
Calculate the standard deviation of zero inflated count data
taking weights for structural zeros into account.
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
sd(Zi)
}
