% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{subset_sample}
\alias{subset_sample}
\title{Subset a \code{\linkS4class{Zi}}-class object based on sample data}
\usage{
subset_sample(Zi, ...)
}
\arguments{
\item{Zi}{\code{\linkS4class{Zi}}-class object}

\item{...}{The subsetting expression that should be applied, see
\link[base]{subset} for more details}
}
\value{
a \code{\linkS4class{Zi}}-class object after subsetting is done
}
\description{
Subset a \code{\linkS4class{Zi}}-class object based on
sample_data of an phyloseq object or on colData based on a
SummarizedExperiment object
}
