% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{t}
\alias{t}
\alias{t,Zi-method}
\title{Transpose a \code{\linkS4class{Zi}}-class object}
\usage{
\S4method{t}{Zi}(x)
}
\arguments{
\item{x}{\code{\linkS4class{Zi}}-class object}
}
\value{
\code{\linkS4class{Zi}}-class object
}
\description{
transpose all matrizes of a \code{\linkS4class{Zi}}-class object
}
