% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{tax_table}
\alias{tax_table}
\alias{tax_table,Zi-method}
\title{Access the taxonomy table}
\usage{
\S4method{tax_table}{Zi}(object)
}
\arguments{
\item{object}{\code{\linkS4class{Zi}}-class object}
}
\value{
tax_table
}
\description{
access the taxonomy table (\link[phyloseq]{tax_table}) of an \code{\linkS4class{Zi}}-class object if the
inputdata slot is a phyloseq object
}
