% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{var}
\alias{var}
\alias{var,Zi,ANY-method}
\title{Variance of zero inflated count data}
\usage{
\S4method{var}{Zi,ANY}(x, na.rm = FALSE)
}
\arguments{
\item{x}{A \code{\linkS4class{Zi}}-class object}

\item{na.rm}{\code{\link[base]{logical}} If \code{\link{TRUE}} \code{\link{NA}}s
are excluded, otherwise not. default = \code{\link{FALSE}}}
}
\value{
variance value
}
\description{
Calculate the variance of zero inflated count data taking weights
for structural zeros into account.
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
var(Zi)
}
\seealso{
\link[matrixStats]{weightedVar}, \link[zitools]{rowVars}, \link[zitools]{colVars}
}
