% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherited_functions.R
\name{weightedVar}
\alias{weightedVar}
\title{Weighted Variance and weighted Standard Deviation}
\usage{
weightedVar(x, w = NULL, idxs = NULL, na.rm = FALSE, center = NULL, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{Zi}}-class object}

\item{w}{a vector of weights the same length as x giving the weights to use for
each element of x}
}
\value{
a numeric scalar
}
\description{
Calculate a weighted variance of zero inflated count data,
additionally taking weights for structural zeros into account
}
\examples{
data(mtx)
Zi <- ziMain(mtx)
weight <- runif(length(Zi@countmatrix), 0.1, 1)
weighted.mean(Zi, w= weight)
}
