% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zi_function.R
\docType{class}
\name{Zi-class}
\alias{Zi-class}
\title{Class Zi}
\description{
Objects of this class store all the results of the ZiMain function to continue
zero inflated data analysis
}
\section{Slots}{

\describe{
\item{\code{inputdata}}{a matrix, phyloseq or SummarizedExperiment object.}

\item{\code{countmatrix}}{matrix. The count matrix, features as rows, samples as columns}

\item{\code{ZiModel}}{list. The result of fitting a zero inflated model using
pscl::zeroinfl}

\item{\code{output}}{matrix. The matrix where predicted structural zeros are omitted
and stored as NA values}

\item{\code{weights}}{matrix. A matrix containing weights for zero counts}
}}

