% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zi_function.R
\name{ziMain}
\alias{ziMain}
\title{ziMain}
\usage{
ziMain(
  datafile,
  feature = "feature",
  formula = count ~ sample + feature,
  dist = "negbin",
  link = "logit",
  zeroRows.rm = FALSE,
  ...
)
}
\arguments{
\item{datafile}{phyloseq object, SummarizedExperiment object, or matrix (
rows=features, columns=samples)}

\item{feature}{"gene", "OTU", "phylum", etc.}

\item{formula}{formula to fit the zero inflated model
response ~ predictor1 + predictor2 + ..., default = count ~ sample+OTU}

\item{dist}{= distribution, either poisson, negative binomial = negbin or geometric}

\item{link}{= link function, either logit, probit, cloglog, cauchit}

\item{zeroRows.rm}{= logical, TRUE if rows that only contain zeros should be removed
or not (they are removed to fit a zero inflated model and will be added afterwards
count matrix per default = 0 and weights = 1)}
}
\value{
S4 list with slots for the input object, the extracted count matrix if
the input is not a matrix, the results of the fitted zero inflation model, a
matrix where the predicted structural zeros are replaced with NA, a matrix
containing the calculated weights
}
\description{
The ziMain function uses a matrix, phyloseq, or SummarizedExperiment
object, extracts the count matrix to fit a zero inflation model to the data.
The matrix is divided into blocks of around 5000 count values to improve run time.
Predicted probabilities given that a zero in the count matrix  is a
structural zero are used to draw structural zeros and replace them with NA. Further,
weights for all zero counts are calculated given the following formula:
w=...
}
\examples{
zero inflated matrix
n <- 1000
male <- sample(c(0,1), size = n, replace = TRUE)
z <- rbinom(n = n, size = 1, prob = 0.3)
mean(z == 0)
y_sim <- ifelse(z == 0, 0,
               rnbinom(n = n,
                       mu = exp(1.3 + 1.5 * (male == 1)),
                       size = 2))
mtx <- matrix(y_sim, 100, 10)


}
